+++ script/tool/lib/src/drive_examples_command.dart
@@ -123,9 +123,6 @@ class DriveExamplesCommand extends PackageLoopingCommand {
 
   @override
   Future<PackageResult> runForPackage(RepositoryPackage package) async {
-    if (!package.requiresFlutter()) {
-      return PackageResult.skip('Not a Flutter package.');
-    }
     final bool isPlugin = isFlutterPlugin(package);
 
     if (package.isPlatformInterface &&