+++ script/tool/test/common/repository_package_test.dart
@@ -96,7 +96,7 @@ void main() {
   });
 
   group('getExamples', () {
-    test('handles a single example', () async {
+    test('handles a single Flutter example', () async {
       final Directory plugin = createFakePlugin('a_plugin', packagesDir);
 
       final List<RepositoryPackage> examples =
@@ -106,7 +106,7 @@ void main() {
       expect(examples[0].path, plugin.childDirectory('example').path);
     });
 
-    test('handles multiple examples', () async {
+    test('handles multiple Flutter examples', () async {
       final Directory plugin = createFakePlugin('a_plugin', packagesDir,
           examples: <String>['example1', 'example2']);
 
@@ -118,6 +118,30 @@ void main() {
           plugin.childDirectory('example').childDirectory('example1').path);
       expect(examples[1].path,
           plugin.childDirectory('example').childDirectory('example2').path);
+    });
+
+    test('handles a single non-Flutter example', () async {
+      final Directory package = createFakePackage('a_package', packagesDir);
+
+      final List<RepositoryPackage> examples =
+          RepositoryPackage(package).getExamples().toList();
+
+      expect(examples.length, 1);
+      expect(examples[0].path, package.childDirectory('example').path);
+    });
+
+    test('handles multiple non-Flutter examples', () async {
+      final Directory package = createFakePackage('a_package', packagesDir,
+          examples: <String>['example1', 'example2']);
+
+      final List<RepositoryPackage> examples =
+          RepositoryPackage(package).getExamples().toList();
+
+      expect(examples.length, 2);
+      expect(examples[0].path,
+          package.childDirectory('example').childDirectory('example1').path);
+      expect(examples[1].path,
+          package.childDirectory('example').childDirectory('example2').path);
     });
   });