+++ script/tool/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## NEXT
 
 - `drive-examples` now supports non-plugin packages.
+- Commands that iterate over examples now include non-Flutter example packages.
 
 ## 0.8.4